requirements.txt
pandas
numpy
scikit-learn
xgboost
lightgbm
click
matplotlib
seaborn
mlflow
jupyter
kagglehub
shap
dvc
hydra-core
optuna